+++ extra_laml_tests.py
@@ -3,7 +3,7 @@
 ##from laml_unit_tests.unit_tests_Simulator import *
 from laml_unit_tests.unit_tests_TopoSearch import *
 from laml_unit_tests.unit_tests_TopoSearchParallel import *
-from laml_unit_tests.unit_tests_io import *
+#from laml_unit_tests.unit_tests_io import *
 import sys
 import os