+++ laml_libs/Topology_search.py
@@ -90,9 +90,11 @@ class Topology_search:
                 if self.has_polytomy:
                     trees,score,params = self.__search_one__(strategy,maxiter=maxiter,verbose=verbose,only_marked=True,checkpoint_file=checkpoint_file)
                 else: # score this tree topology (optimize all numerical params)
+                    if verbose:
+                        print("Found no polytomy to resolve. Optimizing numerical parameters without further topology search")
                     mySolver = self.get_solver()
                     score_tree_strategy = deepcopy(strategy)
-                    score_tree_strategy['fixed_brlen'] = {}
+                    score_tree_strategy['fixed_brlen'] = None
                     score,status = mySolver.score_tree(strategy=score_tree_strategy)
                     self.update_from_solver(mySolver)
                     trees = self.treeTopoList
@@ -100,6 +102,8 @@ class Topology_search:
             else:    
                 if verbose:
                     print("Perform nni moves for full topology search")
+                    if self.has_polytomy:
+                        print("Found polytomies in the input tree(s). Arbitrarily resolving them to obtain fully resolved initial tree(s).") 
                 trees,score,params = self.__search_one__(strategy,maxiter=maxiter,verbose=verbose,only_marked=False,checkpoint_file=checkpoint_file)
             # The final optimization of parameters
             if verbose:
@@ -109,7 +113,7 @@ class Topology_search:
             self.__renew_treeList_obj__()
             mySolver = self.get_solver()
             score_tree_strategy = deepcopy(strategy)
-            score_tree_strategy['fixed_brlen'] = {}
+            score_tree_strategy['fixed_brlen'] = None
             score,status = mySolver.score_tree(strategy=score_tree_strategy)
             self.update_from_solver(mySolver)
             trees = self.treeTopoList
@@ -132,10 +136,13 @@ class Topology_search:
         # optimize branch lengths and other parameters for the starting tree
         mySolver = self.get_solver()
         score_tree_strategy = deepcopy(strategy)
-        score_tree_strategy['fixed_brlen'] = {}
+        score_tree_strategy['fixed_brlen'] = None
         curr_score,status = mySolver.score_tree(strategy=score_tree_strategy)
         if verbose:
-            print("Initial score: " + str(curr_score))
+            if self.has_polytomy:
+                print("Initial score (polytomies were arbitrarily resolved): " + str(curr_score))
+            else:
+                print("Initial score: " + str(curr_score))
         self.update_from_solver(mySolver)
         best_score = curr_score
         best_trees = self.treeTopoList
@@ -199,14 +206,14 @@ class Topology_search:
         # shuffle the order of the nni moves
         shuffle(u_children)
         score_tree_strategy = deepcopy(strategy)
-        score_tree_strategy['fixed_brlen'] = {}
+        score_tree_strategy['fixed_brlen'] = None
 
         if strategy['local_brlen_opt']:
             score_tree_strategy['fixed_nu'] = self.params['nu'] 
             score_tree_strategy['fixed_phi'] = self.params['phi'] 
             free_branches = set(u.child_nodes() + v.child_nodes() + [v])
-            fixed_branches_anchors = []
-            for tree in self.treeList_obj:
+            fixed_branches_anchors = [[] for _ in range(len(self.treeList_obj))]
+            for t,tree in enumerate(self.treeList_obj):
                 for node in tree.traverse_postorder():
                     if node.is_leaf():
                         node.anchors = (node.label,node.label)
@@ -216,14 +223,15 @@ class Topology_search:
                         b = C[-1].anchors[0]
                         node.anchors = (a,b)
                     if not node in free_branches:
-                        fixed_branches_anchors.append(node.anchors)
-            fixed_branches = []
-            for treeTopo in self.treeTopoList:
+                        fixed_branches_anchors[t].append(node.anchors)
+            fixed_branches = [[] for _ in range(len(self.treeList_obj))]
+            for t,treeTopo in enumerate(self.treeTopoList):
                 tree = read_tree_newick(treeTopo)
-                fixed_branches += find_LCAs(tree,fixed_branches_anchors)
-            fixed_brlen = {}
-            for i,node in enumerate(fixed_branches):
-                fixed_brlen[fixed_branches_anchors[i]] = node.edge_length
+                fixed_branches[t] += find_LCAs(tree,fixed_branches_anchors[t])
+            fixed_brlen = [{} for _ in range(len(self.treeList_obj))]
+            for t,B in enumerate(fixed_branches):
+                for i,node in enumerate(B):
+                    fixed_brlen[t][fixed_branches_anchors[t][i]] = node.edge_length
             score_tree_strategy['fixed_brlen'] = fixed_brlen
 
         for u_child in u_children:
@@ -238,7 +246,7 @@ class Topology_search:
             mySolver = self.solver([tree.newick() for tree in self.treeList_obj],self.data,self.prior,self.params)            
             new_score,status = mySolver.score_tree(strategy=score_tree_strategy)
             if status != "optimal" and strategy['local_brlen_opt']:
-                score_tree_strategy['fixed_brlen'] = {}
+                score_tree_strategy['fixed_brlen'] = None
                 mySolver = self.solver([tree.newick() for tree in self.treeList_obj],self.data,self.prior,self.params)            
                 new_score,status = mySolver.score_tree(strategy=score_tree_strategy)
             if self.__accept_proposal__(curr_score,new_score,nni_iter): # accept the new tree and params