+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 * LAML version 0.0.4
     * handle mismatch between character matrix and prior pickle
+    * fix error in doing topology search on multiple input trees
+    * more informative messages on handling polytomies
 * LAML version 0.0.3 (depricated)
     * attempt to allow user-specified sampling times
     * seems to work for branch length estimation on a fixed topology, but fails on doing topology search